- Fix card modal not opening at first click on hall of fame - Refactored a function to parse the deck cards info
