+++ react-frontend/src/components/twd_hall_fame/index.js
@@ -1,0 +1,3 @@
+export { default as TwdHallFameDeckBody } from './TwdHallFameDeckBody.jsx';
+export { default as TwdHallFameDeckHeader } from './TwdHallFameDeckHeader.jsx';
+export { default as TwdHallFamePlayer } from './TwdHallFamePlayer.jsx';
